% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_blob_column}
\alias{add_blob_column}
\title{Add blob column}
\usage{
add_blob_column(column_name, table_name, conn)
}
\arguments{
\item{column_name}{A string of the name of the column.}

\item{table_name}{A string of the name of the table.}

\item{conn}{A connection object.}
}
\value{
Modified SQLite database.
}
\description{
Add named empty blob column to SQLite database
}
