% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflob.R
\name{blob_column}
\alias{blob_column}
\title{Add blob column}
\usage{
blob_column(table_name, column_name, conn)
}
\arguments{
\item{table_name}{A string of the name of the table.}

\item{column_name}{A string of the name of the column.}

\item{conn}{A connection object.}
}
\value{
Modified SQLite database.
}
\description{
Add named empty blob column to SQLite database
}
