% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflobr-package.R
\docType{package}
\name{dbflobr-package}
\alias{dbflobr}
\alias{dbflobr-package}
\title{dbflobr: Read and Write Files to  SQLite Databases as BLOBs}
\description{
Reads and writes files to SQLite databases as flobs 
    (a special type of BLOB that includes the file extension type).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dbflobr}
  \item Report bugs at \url{https://github.com/poissonconsulting/dbflobr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Dalgarno \email{seb@poissonconsulting.ca}

Other contributors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592) [contributor]
}

}
\keyword{internal}
