% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflob.R
\name{delete_flob}
\alias{delete_flob}
\title{Delete blob}
\usage{
delete_flob(column_name, table_name, key, conn)
}
\arguments{
\item{column_name}{A string of the name of the BLOB column.}

\item{table_name}{A string of the name of the table.}

\item{key}{A data.frame with column names and values which filter table to a single row.}

\item{conn}{A connection object.}
}
\value{
A flob.
}
\description{
Delete a blob from SQLite database.
}
