% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflob.R
\name{flob_collapse}
\alias{flob_collapse}
\title{Collapse flob}
\usage{
flob_collapse(x)
}
\arguments{
\item{x}{A flob.}
}
\value{
Modified SQLite database.
}
\description{
Collapse flob into format readable by SQLite
}
