% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_flobs}
\alias{import_flobs}
\title{Import flobs.}
\usage{
import_flobs(
  column_name,
  table_name,
  key,
  conn,
  dir = ".",
  exists = FALSE,
  recursive = FALSE
)
}
\arguments{
\item{column_name}{A string of the name of the BLOB column.}

\item{table_name}{A string of the name of the existing table.}

\item{key}{An empty data.frame whose columns and types are populated by
file names in dir. The populated key is used to filter the table to a
single row (this in combination with the column_name argument are used to
target a single cell within the table to modify).}

\item{conn}{A SQLite connection object.}

\item{dir}{A string of the path to the directory to save the files in.}

\item{exists}{A logical scalar specifying whether the column must (TRUE) or
mustn't (FALSE) already exist or whether it doesn't matter (NA).
IF FALSE, a new BLOB column is created.}

\item{recursive}{A flag indicating whether to recurse into file directory.}
}
\value{
An invisible named vector of file names and flag indicating whether
successfully written to database.
}
\description{
Import flobs to SQLite database from directory.
}
\examples{
flob <- flobr::flob_obj
conn <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
DBI::dbGetQuery(conn, "CREATE TABLE Table1 (IntColumn INTEGER PRIMARY KEY NOT NULL)")
DBI::dbWriteTable(conn, "Table1", data.frame(IntColumn = c(1L, 2L)), append = TRUE)
key <- data.frame(IntColumn = 2L)
write_flob(flob, "BlobColumn", "Table1", key, conn, exists = FALSE)
dir <- tempdir()
save_flobs("BlobColumn", "Table1", conn, dir)
DBI::dbDisconnect(conn)
}
