% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflob.R
\name{read_flob}
\alias{read_flob}
\title{Read flob}
\usage{
read_flob(column_name, table_name, key, conn)
}
\arguments{
\item{column_name}{A string of the name of the BLOB column.}

\item{table_name}{A string of the name of the table.}

\item{key}{A data.frame with column names and values which filter table to a single row.}

\item{conn}{A connection object.}
}
\value{
A flob.
}
\description{
Read a blob from SQLite database into flob.
}
