% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_all_flobs}
\alias{save_all_flobs}
\title{Save all flobs.}
\usage{
save_all_flobs(table_name = NULL, dir, conn)
}
\arguments{
\item{table_name}{A vector of character strings indicating names of tables to save flobs from.
By default all tables are included.}

\item{dir}{A character string of the path to the directory to save files to.}

\item{conn}{A SQLite connection object.}
}
\value{
An invisible character string of the directory.
}
\description{
Rename \code{\link[flobr]{flob}}s from a SQLite database and save to directory.
}
\examples{
flob <- flobr::flob_obj
conn <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
DBI::dbWriteTable(conn, "Table1", data.frame(IntColumn = c(1L, 2L)))
key <- data.frame(IntColumn = 2L)
write_flob(flob, "BlobColumn", "Table1", key, conn, exists = FALSE)
dir <- tempdir()
save_all_flobs(dir = dir, conn = conn)
DBI::dbDisconnect(conn)
}
