% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflob.R
\name{write_flob}
\alias{write_flob}
\title{Write flob}
\usage{
write_flob(flob, conn, table_name, column_name, rowid)
}
\arguments{
\item{flob}{A flob.}

\item{conn}{A connection object.}

\item{table_name}{A string of the name of the table.}

\item{column_name}{A string of the name of the column.}

\item{rowid}{An integer of the table rowid.}
}
\value{
Modified database.
}
\description{
Write a flob to column of type BLOB in database.
}
