% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflobr-package.R
\docType{package}
\name{dbflobr-package}
\alias{dbflobr}
\alias{dbflobr-package}
\title{dbflobr: Read and Write Files to SQLite Databases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reads and writes files to SQLite databases \url{https://www.sqlite.org/index.html} as flobs (a flob is a blob that preserves the file extension).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dbflobr}
  \item Report bugs at \url{https://github.com/poissonconsulting/dbflobr/issues}
}

}
\author{
\strong{Maintainer}: Evan Amies-Galonski \email{evan@poissonconsulting.ca} (\href{https://orcid.org/0000-0003-1096-2089}{ORCID})

Authors:
\itemize{
  \item Sebastian Dalgarno \email{seb@northbeachconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
