% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflobr-package.R
\docType{package}
\name{dbflobr-package}
\alias{dbflobr}
\alias{dbflobr-package}
\title{dbflobr: Read and Write Files to SQLite Databases}
\description{
Reads and writes files to SQLite databases <https://www.sqlite.org/index.html> as flobs 
    (a flob is a blob that preserves the file extension).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dbflobr}
  \item Report bugs at \url{https://github.com/poissonconsulting/dbflobr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Dalgarno \email{seb@poissonconsulting.ca} (0000-0002-3658-4517)

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)
}

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
