% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbflob.R
\name{read_flob}
\alias{read_flob}
\title{Read flob}
\usage{
read_flob(conn, table_name, column_name, rowid)
}
\arguments{
\item{conn}{An connection object.}

\item{table_name}{A string of the name of the table.}

\item{column_name}{A string of the name of the column.}

\item{rowid}{An integer of the table rowid.}
}
\value{
A flob.
}
\description{
Read a blob from SQLite database into flob.
}
